Promote tmux fallback into a real plugin
Step 5 of the backend-plugin refactor: the tmux fallback transport now
lives in `backends/tmux/` with the same shape as the Claude and Codex
plugins. Capabilities advertise `is_structured=False` so the frontend
keeps rendering the heuristic transcript, `supports_resume=True` for
re-attaching detached panes, and disables every inline control knob
since no protocol path exists to set them mid-session.

`_legacy_shims.py` collapses to the default registry assembler now
that no shim classes remain.
