+++ backend/src/waypoint/backends/tmux/__init__.py
@@ -1,0 +1,3 @@
+from waypoint.backends.tmux.plugin import TmuxPlugin, build_plugin
+
+__all__ = ["TmuxPlugin", "build_plugin"]
